+++ gogod/sms.py
@@ -1,7 +1,12 @@
 import commands
 import threading
 import gammu
+import os
+import consolelog
 
+GAMMU_CONFIG_FILE   = "/root/.gammurc"
+USB_CONFIG_FILE     = "/etc/usb_modeswitch.conf"
+LOG_TITLE           = "SMS"
 
 class SmsStatus(object):
     NEW                 = 0
@@ -21,10 +26,49 @@ class SmsHandle(threading.Thread):
         self.sms_number = sms_number
         self.sms_message = sms_message
 
+
+    def config(self):
+        if not os.path.exists(GAMMU_CONFIG_FILE):
+            consolelog.log(LOG_TITLE, "No gammu config File")
+            gammu_config_content = '''[gammu]
+
+port = /dev/ttyUSB0
+model =
+connection = at19200
+synchronizetime = yes
+logfile =
+logformat = nothing
+use_locking =
+gammuloc =
+            '''
+
+            file = open(GAMMU_CONFIG_FILE, "w+")
+            file.write(gammu_config_content)
+            file.close()
+            consolelog.log(LOG_TITLE, "Written config to file")
+
+
+        if 'DefaultVendor' not in open(USB_CONFIG_FILE).read():
+            consolelog.log(LOG_TITLE, "No Device config")
+            modeswitch_content = '''DefaultVendor= 0x12d1
+DefaultProduct = 0x1446
+TargetVendor = 0x12d1
+TargetProductList ="1001,1406,140b,140c,141b,14ac"
+CheckSuccess = 20
+MessageContent ="55534243123456780000000000000011060000000000000000000000000000"'''
+
+            file = open(USB_CONFIG_FILE, 'a')
+            file.write(modeswitch_content)
+            file.close()
+            consolelog.log(LOG_TITLE, "Written device config")
+
+        os.system("sudo usb_modeswitch -c /etc/usb_modeswitch.conf")
+
     def run(self):
         self.status = SmsStatus.SENDING
         self.status_callback(self.status)
-        print "SMS : thread started"
+        consolelog.log(LOG_TITLE, "thread started")
+        self.config()
         try:
             #command_string = 'sudo gammu sendsms TEXT ' + self.sms_number + ' -textutf8 "' + self.sms_message + '" '
             #print commands.getoutput(command_string)
@@ -39,16 +83,16 @@ class SmsHandle(threading.Thread):
             }
             # Actually send the message
             result = sm.SendSMS(message)
-            print("SMS : Send sms Success")
+            consolelog.log(LOG_TITLE, "Send sms Success")
             self.status = SmsStatus.SUCCESS
         except gammu.ERR_DEVICENOTEXIST:
-            print "SMS : Error device doesn't exist"
+            consolelog.log(LOG_TITLE, "Error device doesn't exist")
             self.status = SmsStatus.ERR_DEVICENOTEXIST
         except gammu.ERR_NOSIM:
-            print "SMS : Error can not access SIM card"
+            consolelog.log(LOG_TITLE, "Error can not access SIM card")
             self.status = SmsStatus.ERR_NOSIM
         except ValueError:
-            print "SMS : error"
+            consolelog.log(LOG_TITLE, "Error")
             self.status = SmsStatus.ERROR
         self.status_callback(self.status)
 
@@ -79,7 +123,7 @@ def read_all_sms():
 
 
 def send(status_callback, sms_number, sms_message):
-    print "SMS : starting sms thread"
+    consolelog.log(LOG_TITLE, "starting sms thread")
     sms_thread = SmsHandle(status_callback, sms_number, sms_message)
     sms_thread.start()